api/pyproject.toml
[tool.poetry]
name = "api"
version = "1.0.0"
description = ""
authors = ["Adrian Romo edgar.adrian97@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
djangorestframework-simplejwt = "^4.6.0"
django-cors-headers = "^3.7.0"
psycopg2-binary = "^2.8.4"
drf-yasg = "^1.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
